// apps/market-monitor-cloud-functions/src/model/firebase-stock.model.ts
export enum StockDataFields {
	// collection names
	stock_data = 'stock_data',
	more_information = 'more_information',

	// document names
	stock_details = 'stock_details',
	stock_historical_1d = 'stock_historical_1d',
	stock_historical_1wk = 'stock_historical_1wk',
	stock_historical_1mo = 'stock_historical_1mo',
	stock_historical_3mo = 'stock_historical_3mo',
	stock_historical_6mo = 'stock_historical_6mo',
	stock_historical_ytd = 'stock_historical_ytd',
	stock_historical_1yr = 'stock_historical_1yr',
	stock_historical_5yr = 'stock_historical_5yr',
	stock_historical_all = 'stock_historical_all',
	stock_metrics = 'stock_metrics',
	stock_ratios = 'stock_ratios',
	stock_financial_statements = 'stock_financial_statements',
	stock_financial_growth = 'stock_financial_growth',
}